+++ R/pn.R
@@ -1,0 +1,13 @@
+#' Print the current time friendly
+#'
+#' @return Character for print
+#' @import lubridate
+#' @export
+#' @examples
+#' pn()
+
+pn <- function(){
+    lubridate::now() %>%
+        as.character() %>%
+        cat
+}